Re-organise tests to reflect updated package structure
